agents/preprocessor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
pillow==10.1.0
librosa==0.10.1
opencv-python-headless==4.8.1.78
pydantic==2.5.0
httpx==0.25.2
structlog==23.2.0